Update travis to build dc-core itself
For linux this builds does everything in a docker image.  This avoids
having to re-build all the dependencies on each CI build.  In turn
this is useful because the travis linux OS is somewhat old and we'd
have to build a number of dependencies from source.

This also introduces a matrix and builds for:
- os & linux
- node.js 10
- deltachat-core latest release & HEAD