Use installed libdeltachat.so (#257)
* Use installed libdeltachat.so

This changes the bindings to build against the installed version of
libdeltachat.so using pkg-config.

It is possible to use an un-installed version by setting a few
environment variables.

* Update travis to build dc-core itself

For linux this builds does everything in a docker image.  This avoids
having to re-build all the dependencies on each CI build.  In turn
this is useful because the travis linux OS is somewhat old and we'd
have to build a number of dependencies from source.

This also introduces a matrix and builds for:
- os & linux
- node.js 10
- deltachat-core latest release & HEAD

* Simpler way to build against uninstalled dc-core

This majorly simplifies the way to build against an uninstalled
libdeltachat.so and cleans up the bindings.gyp file.

* Build zlib from source

It seems there's a bug in node/npm which gets triggered by an older
zlib version.

We also like to have the multiarch lib directory in ld.so.conf.

* Run ldconfig and use host passwd file

After installing dc-core we need to run ldconfig.

Start the docker container with the host's /etc/passwd file so that
the user ID matches and is known.

* Eventqueue is no more

Fixing up a bad rebase.

* Move to ci_scripts to unify naming across repos

* We're compiling deltachat-core again!

- binding.gyp by default now links to bundled deltachat-core
- readded rebuilding scripts for core & readded submodule too
- changed way of enabling system installd libdeltachat behind --dc-stystem-lib=true
- updated Readme

* Fix standard