Use externally-valid include again
This switches back to using #include <deltachat/deltachat.h> which is
the way the header should be included for users of libdeltachat.  This
needed
https://github.com/deltachat/deltachat-core/commit/12ef73c8e76185f9b78e844ea673025f56a959ab
which was not yet committed when this was first made.

It allows us to include things the correct way both when compiling
against an installed and against an un-installed library without
having to abuse the -I flags.